refactor: rename vars, delete global vars
