+++ random-game/scripts/helpers/localStorage.js
